requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyjwt==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
cryptography==42.0.5
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Document parsing dependencies
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0








